test: move editor e2e tests to a separate module
