+++ packages/cli-repl/test/e2e.spec.ts
@@ -9,7 +9,7 @@ import { promisify } from 'util';
 import rimraf from 'rimraf';
 import path from 'path';
 import os from 'os';
-import { readReplLogfile, fakeExternalEditor, setTemporaryHomeDirectory } from './repl-helpers';
+import { readReplLogfile, setTemporaryHomeDirectory } from './repl-helpers';
 import { bson } from '@mongosh/service-provider-core';
 const { EJSON } = bson;
 
@@ -790,8 +790,6 @@ describe('e2e', function() {
       homedir = homeInfo.homedir;
       env = homeInfo.env;
       if (process.platform === 'win32') {
-        env.LOCALAPPDATA = path.join(homedir, 'local');
-        env.APPDATA = path.join(homedir, 'roaming');
         logBasePath = path.resolve(homedir, 'local', 'mongodb', 'mongosh');
         configPath = path.resolve(homedir, 'roaming', 'mongodb', 'mongosh', 'config');
         historyPath = path.resolve(homedir, 'roaming', 'mongodb', 'mongosh', 'mongosh_repl_history');
@@ -1140,121 +1138,6 @@ describe('e2e', function() {
       });
       expect(await shell.waitForExit()).to.equal(0);
       shell.assertContainsOutput('610');
-    });
-  });
-
-  describe('external editor', () => {
-    let db;
-    let client;
-    let shell: TestShell;
-    let dbName;
-    let homedir: string;
-    let env: Record<string, string>;
-
-    beforeEach(async() => {
-      const homeInfo = setTemporaryHomeDirectory();
-      const connectionString = await testServer.connectionString();
-
-      dbName = `test-${Date.now()}`;
-      homedir = homeInfo.homedir;
-      env = homeInfo.env;
-
-      shell = TestShell.start({
-        args: [ connectionString ],
-        forceTerminal: true,
-        env
-      });
-
-      client = await MongoClient.connect(connectionString, {});
-      db = client.db(dbName);
-
-      await shell.waitForPrompt();
-      shell.assertNoErrors();
-    });
-
-    afterEach(async() => {
-      await db.dropDatabase();
-      client.close();
-      await TestShell.killall.call(this);
-      try {
-        await promisify(rimraf)(homedir);
-      } catch (err) {
-        // On Windows in CI, this can fail with EPERM for some reason.
-        // If it does, just log the error instead of failing all tests.
-        console.error('Could not remove fake home directory:', err);
-      }
-    });
-
-    it('returns a modified identifier for fn', async() => {
-      const shellOriginalInput = "const fn = function () { console.log(111); }; edit('fn')";
-      const editorOutput = `function () {
-        console.log(222);
-      };`;
-      const shellModifiedInput = 'fn = function () { console.log(222); };';
-      const editor = await fakeExternalEditor(editorOutput);
-      const result = await shell.executeLine(`config.set("editor", "${editor}");`);
-
-      expect(result).to.include('"editor" has been changed');
-      shell.writeInputLine(shellOriginalInput);
-      await eventually(() => {
-        shell.assertContainsOutput(shellModifiedInput);
-      });
-    });
-
-    it('returns a modified identifier for var', async() => {
-      const shellOriginalInput = "const myVar = '111'; edit('myVar')";
-      const editorOutput = "const myVar = '222';";
-      const shellModifiedInput = "myVar = '222';";
-      const editor = await fakeExternalEditor(editorOutput);
-      const result = await shell.executeLine(`config.set("editor", "${editor}");`);
-
-      expect(result).to.include('"editor" has been changed');
-      shell.writeInputLine(shellOriginalInput);
-      await eventually(() => {
-        shell.assertContainsOutput(shellModifiedInput);
-      });
-    });
-
-    it('returns a modified identifier for a.b.c', async() => {
-      const shellOriginalInput = "const myObj = { field: { child: 'string value' } }; edit('myObj')";
-      const editorOutput = `const myObj = {
-        field: {
-          child: 'new   value'
-        }
-      };`;
-      const shellModifiedInput = "myObj = { field: { child: 'new   value' } };";
-      const editor = await fakeExternalEditor(editorOutput);
-      const result = await shell.executeLine(`config.set("editor", "${editor}");`);
-
-      expect(result).to.include('"editor" has been changed');
-      shell.writeInputLine(shellOriginalInput);
-      await eventually(() => {
-        shell.assertContainsOutput(shellModifiedInput);
-      });
-    });
-
-    it('returns an error when editor exits with exitCode 1', async() => {
-      const shellOriginalInput = 'edit db.test.foobar';
-      const editor = await fakeExternalEditor();
-      const result = await shell.executeLine(`config.set("editor", "${editor}");`);
-
-      expect(result).to.include('"editor" has been changed');
-      shell.writeInputLine(shellOriginalInput);
-      await eventually(() => {
-        shell.assertContainsError('failed with an exit code 1');
-      });
-    });
-
-    it('opens an empty editor', async() => {
-      const output = '';
-      const editor = await fakeExternalEditor(output);
-      const result = await shell.executeLine(`config.set("editor", "${editor}");`);
-
-      expect(result).to.include('"editor" has been changed');
-      shell.writeInputLine('edit');
-      await eventually(() => {
-        shell.assertContainsOutput(output);
-      });
     });
   });
 });
